Renamed the 'Google' component folder to 'Transcription' for clarity #67
