+++ RC-SpeechToText/ClientApp/components/Transcription/EditTranscriptionButton.tsx
