+++ RC-SpeechToText/ClientApp/components/Transcription/SearchField.tsx
