+++ RC-SpeechToText/ClientApp/components/Transcription/Transcription.tsx
