+++ RC-SpeechToText/ClientApp/components/Transcription/TranscriptionButton.tsx
