+++ RC-SpeechToText/ClientApp/components/Transcription/TranscriptionText.tsx
