+++ RC-SpeechToText/ClientApp/components/Transcription/VideoInput.tsx
